conversor-imagenes/requirements.txt
Pillow>=10.0.0
pillow-avif-plugin>=1.4.0